src/ingest/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
google-cloud-pubsub
requests
